requirements.txt
Django==2.0
django-cors-headers==2.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
mysqlclient==1.3.12
pytz==2017.3
